# refactor/runner/openapi.yaml
openapi: 3.0.2
info:
  title: Runner API
  version: 0.1.0
paths:
  /api/v0/task/run:
    post:
      tags:
        - task
      summary: Run Task
      description: API Endpoint (/api/task/run) to run tasks
      operationId: run_task_api_v0_task_run_post
      parameters:
        - required: true
          schema:
            title: Dispatch Id
            type: string
          name: dispatch_id
          in: query
      requestBody:
        content:
          application/json:
            schema:
              title: Task Ids
              type: array
              items:
                type: integer
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunTaskResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v0/task/status:
    get:
      tags:
        - task
      summary: Check Status
      description: API Endpoint (/api/task/status) to check status of a task
      operationId: check_status_api_v0_task_status_get
      parameters:
        - required: true
          schema:
            title: Dispatch Id
            type: string
          name: dispatch_id
          in: query
        - required: true
          schema:
            title: Task Id
            type: integer
          name: task_id
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v0/task/cancel:
    post:
      tags:
        - task
      summary: Cancel Task
      description: API Endpoint (/api/task/cancel) to cancel a task
      operationId: cancel_task_api_v0_task_cancel_post
      parameters:
        - required: true
          schema:
            title: Dispatch Id
            type: string
          name: dispatch_id
          in: query
        - required: true
          schema:
            title: Task Id
            type: integer
          name: task_id
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    CancelResponse:
      title: CancelResponse
      required:
        - cancelled_dispatch_id
        - cancelled_task_id
      type: object
      properties:
        cancelled_dispatch_id:
          title: Cancelled Dispatch Id
          type: string
        cancelled_task_id:
          title: Cancelled Task Id
          type: string
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    RunTaskResponse:
      title: RunTaskResponse
      required:
        - response
      type: object
      properties:
        response:
          title: Response
          type: string
    TaskStatus:
      title: TaskStatus
      required:
        - status
      type: object
      properties:
        status:
          title: Status
          type: string
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string